pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "localforge"
version = "0.1.0"
description = "Local AI coding station — MCP server + multi-device mesh + autonomous agents"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
authors = [
    { name = "BitwiseBard" },
]
keywords = ["mcp", "llm", "local-ai", "agents", "rag", "code-review"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

# Core dependencies — required for MCP server to function
dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "tenacity>=9.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
# HTTP gateway (dashboard, remote access, CLI)
gateway = [
    "uvicorn>=0.30.0",
    "starlette>=0.38.0",
]

# Semantic search (dense + sparse + ColBERT + reranker)
embeddings = [
    "fastembed>=0.3.0",
    "numpy>=1.24.0",
]

# Code-aware chunking for RAG indexing
search = [
    "tree-sitter-languages>=1.10.0",
]

# Autonomous agent file-watch triggers
agents = [
    "watchdog>=6.0.0",
]

# Web search and research tools
web = [
    "ddgs>=2.0.0",
    "trafilatura>=2.0.0",
]

# Media processing (photo/video)
media = [
    "Pillow>=10.0.0",
]

# All optional features
all = [
    "localforge[gateway,embeddings,search,agents,web,media]",
]

# Development tools
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[project.urls]
Homepage = "https://github.com/bitwisebard/localforge"
Repository = "https://github.com/bitwisebard/localforge"
Issues = "https://github.com/bitwisebard/localforge/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/localforge"]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
